ProcessMonitor: Update
In this version, the method focuses on the vrserver process. It checks if this process was started by your application and has exited. If it has exited with an ExitCode of 0, it's considered a normal, user-initiated exit. If the process hasn't exited, or if it exited with a non-zero exit code, it's considered an abnormal exit.